refactor(nft): Move routes inside index
